Extracted ConnectWallet into its own compone
